+++ migrations/20250324173525_update_history.sql
@@ -1,8 +1,8 @@
 -- Add migration script here
 CREATE TABLE update_history (
-    'time' DATETIME NOT NULL,
+    'date' DATETIME NOT NULL,
     'name' VARCHAR(256) NOT NULL,
-    'current_version' VARCHAR(256) NOT NULL,
+    'old_version' VARCHAR(256) NOT NULL,
     'updated_to' VARCHAR(256) NOT NULL,
-    PRIMARY KEY ('time') 
+    PRIMARY KEY ('date') 
 );