+++ migrations/20250324203354_update_checks_add_updates_available.sql
@@ -1,0 +1,9 @@
+-- Add migration script here
+ALTER TABLE update_checks
+RENAME TO update_check_history;
+
+ALTER TABLE update_check_history
+ADD updates_available INTEGER;
+
+ALTER TABLE update_check_history
+ADD programs VARCHAR(4096);