+++ changelog.md
@@ -12,7 +12,7 @@
           Set this flag to not mark the update as seen and to make the notification get sent when run-timed mode is used the next time.
 ```
 
-- performed updates are now stored in the database
+- performed updates are now stored in the database, they can be shown using the new `update-history` subcommand
 
 ## v1.2.1